% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras.R
\name{barras}
\alias{barras}
\title{Graficos de barras do tipo highcharter}
\usage{
barras(da, tipo = "bar")
}
\arguments{
\item{da}{base}

\item{tipo}{tipo barras}
}
\description{
funcao que cria grafico dinamico
}
