% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_empilhadas.R
\name{barras_empilhadas}
\alias{barras_empilhadas}
\title{Grafico de colunas do tipo highcharter por grupo}
\usage{
barras_empilhadas(base, grupo, tipo = "column")
}
\arguments{
\item{base}{base que gera o grafico}

\item{grupo}{grupo de informacaoes}

\item{tipo}{column}
}
\description{
funcao que cria grafico dinamico de colunas por grupo
}
