% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_pgfn.R
\name{barras_pgfn}
\alias{barras_pgfn}
\title{Graficos de barras do tipo highcharter}
\usage{
barras_pgfn(da, tipo = "bar", valor, corte = 0.01)
}
\arguments{
\item{da}{base}

\item{tipo}{barras}

\item{valor}{valor}

\item{corte}{intervalo (default é 0.01)}
}
\description{
funcao que cria grafico dinamico de barras com proporção
}
