% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_prop.R
\name{barras_prop}
\alias{barras_prop}
\title{Graficos de barras do tipo highcharter}
\usage{
barras_prop(da)
}
\arguments{
\item{da}{base}
}
\description{
funcao que cria grafico dinamico de barras com proporção
}
