% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{base_teste}
\alias{base_teste}
\title{base_teste}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 335 rows and 5 columns.
}
\usage{
base_teste
}
\description{
base_teste
}
\keyword{datasets}
