% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_estatisticas.R
\name{calcular_estatisticas}
\alias{calcular_estatisticas}
\title{Calcula estatísticas de tempo}
\usage{
calcular_estatisticas(intervalo, da)
}
\arguments{
\item{intervalo}{nome do intervalo}

\item{da}{base de dados}
}
\description{
funcao que ajuda calcular medias e medianas por
intervalo
}
