% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_medianas.R
\name{calcular_medianas}
\alias{calcular_medianas}
\title{Calcula as medianas com base nos dados, datas e quebra}
\usage{
calcular_medianas(da, data1, data2, quebra)
}
\arguments{
\item{da}{base de dados filtrada}

\item{data1}{coluna de data da esquerda}

\item{data2}{coluna de data da direita}

\item{quebra}{quebra de tempo}
}
\description{
funcao que ajuda calcular medianas por quebra de tempo
}
