% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_test.R
\name{conta_itens}
\alias{conta_itens}
\title{teste para testthat contar media}
\usage{
conta_itens(x, compact)
}
\arguments{
\item{x}{vetor}

\item{compact}{(default true)}
}
\description{
funcoo que tira a media de um vetor
}
