% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_median.R
\name{get_median}
\alias{get_median}
\title{Pegar a mediana baseada no processo}
\usage{
get_median(da, re_tempo)
}
\arguments{
\item{da}{dados}

\item{re_tempo}{seletor de tempos}
}
\description{
Função que ajuda pegar a mediana de processos por
tempo
}
