% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n.R
\name{get_n}
\alias{get_n}
\title{Pegar a qtd de casos baseada no processo}
\usage{
get_n(da, re_tempo)
}
\arguments{
\item{da}{dados}

\item{re_tempo}{seletor de tempos}
}
\description{
funcao que ajuda pegar a quantidade de processos por
tempo
}
