% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_km.R
\name{grafico_km}
\alias{grafico_km}
\title{Gráfico Kaplan-meier de modelo de sobrevivência}
\usage{
grafico_km(m)
}
\arguments{
\item{m}{modelo}
}
\description{
Função que cria gráfico no modelo de sobrevivência Kaplan-meier
}
