% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_km2.R
\name{grafico_km2}
\alias{grafico_km2}
\title{Grafico Kaplan-meier de modelo de sobrevivencia (2 variaveis)}
\usage{
grafico_km2(m, lab)
}
\arguments{
\item{m}{modelo}

\item{lab}{label para cores}
}
\description{
funcao que cria grafico no modelo de sobrevivendia Kaplan-meier
}
