% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprop.R
\name{nprop}
\alias{nprop}
\title{Ferramenta para cálculo porcentagem}
\usage{
nprop(coluna, val)
}
\arguments{
\item{coluna}{coluna}

\item{val}{categorias de teste}
}
\description{
Função que retorna no formato NN (PP)
}
