% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_para_cor.R
\name{status_para_cor}
\alias{status_para_cor}
\title{Status cores}
\usage{
status_para_cor(status)
}
\arguments{
\item{status}{info, secondary, primary, success, warning, danger}
}
\description{
funcao que cria cores para os status dos dashboards
}
