% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela.R
\name{tabela}
\alias{tabela}
\title{Formata tabelas com caracteristicas parecidas}
\usage{
tabela(da, ..., .lump = 0, .nm = NULL)
}
\arguments{
\item{da}{base}

\item{...}{outros argumentos}

\item{.lump}{default 0}

\item{.nm}{numeros}
}
\description{
funcoo que ajuda definir tabelas
}
