% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_app_server.R
\name{template_app_server}
\alias{template_app_server}
\title{template_app_server}
\usage{
template_app_server(
  nome_arquivo = "app_server",
  path = "R",
  template = "templates/template_app_server.R"
)
}
\arguments{
\item{nome_arquivo}{nome do arquivo}

\item{path}{nome do diretório}

\item{template}{nome do template que deseja}
}
\description{
Cria templates para o App Server
}
