% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_mod_server.R
\name{template_mod_server}
\alias{template_mod_server}
\title{template_mod_server}
\usage{
template_mod_server(
  nome_arquivo = "mod_processo_server",
  path = "R",
  template = "templates/template_mod_processo_server.R"
)
}
\arguments{
\item{nome_arquivo}{nome do arquivo}

\item{path}{nome do diretório}

\item{template}{nome do template que deseja}
}
\description{
Cria templates para o modulo de Server
}
