% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_mod_ui.R
\name{template_mod_ui}
\alias{template_mod_ui}
\title{template_mod_ui}
\usage{
template_mod_ui(
  nome_arquivo = "mod_processo_ui",
  path = "R",
  template = "templates/template_mod_processo_ui.R"
)
}
\arguments{
\item{nome_arquivo}{nome do arquivo}

\item{path}{nome do diretório}

\item{template}{nome do template que deseja}
}
\description{
Cria templates para o  módulo de UI
}
