% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_lab_one.R
\name{tempo_lab_one}
\alias{tempo_lab_one}
\title{Ferramenta para facilitar o calculo de tempos por ano}
\usage{
tempo_lab_one(numeric_expression)
}
\arguments{
\item{numeric_expression}{coluna}
}
\description{
funcao que ajuda calcular tempos
}
