% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_test.R
\name{tira_media}
\alias{tira_media}
\title{teste para testthat tira media}
\usage{
tira_media(x, rm_na = TRUE)
}
\arguments{
\item{x}{vetor}

\item{rm_na}{(default true)}
}
\description{
funcoo que tira a media de um vetor
}
