% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_empilhadas.R
\name{barras_empilhadas}
\alias{barras_empilhadas}
\title{Graficos de barras do tipo highcharter por grupo}
\usage{
barras_empilhadas(da, grupo, tipo = "column")
}
\arguments{
\item{da}{base}

\item{grupo}{grupo de informacaoes}

\item{tipo}{column}
}
\description{
funcao que cria grafico dinamico
}
