% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_empilhadas.R
\name{barras_empilhadas}
\alias{barras_empilhadas}
\title{Gráfico de colunas do tipo highcharter por grupo}
\usage{
barras_empilhadas(base, grupo, tipo = "column")
}
\arguments{
\item{base}{base que gera o gráfico}

\item{grupo}{grupo de informações}

\item{tipo}{column}
}
\description{
Funcão que cria gráfico dinâmico de colunas por grupo
}
