% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_prop.R
\name{barras_prop}
\alias{barras_prop}
\title{Graficos de barras do tipo highcharter}
\usage{
barras_prop(base)
}
\arguments{
\item{base}{base que gera o grafico}
}
\description{
funcao que cria grafico dinamico de barras com proporcao
}
