% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_estatisticas.R
\name{calcular_estatisticas}
\alias{calcular_estatisticas}
\title{Calcula estatísticas de tempo}
\usage{
calcular_estatisticas(intervalo, base)
}
\arguments{
\item{intervalo}{nome do intervalo}

\item{base}{base de dados}
}
\description{
Função que ajuda calcular médias e medianas por
intervalo
}
