% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_lab.R
\name{custom_lab}
\alias{custom_lab}
\title{Helper para tratar milhares e milh천es}
\usage{
custom_lab(x, ..., dinheiro = TRUE)
}
\arguments{
\item{x}{coluna}

\item{...}{outros argumentos (n찾o utilizado)}

\item{dinheiro}{transformar em dinheiro? Por padr찾o sim}
}
\description{
Helper para tratar milhares e milh천es
}
