% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_lab.R
\name{custom_lab}
\alias{custom_lab}
\title{Helper para tratar milhares e milhões}
\usage{
custom_lab(x, ..., dinheiro = TRUE)
}
\arguments{
\item{x}{coluna}

\item{...}{outros argumentos (não utilizado)}

\item{dinheiro}{transformar em dinheiro? Por padrao sim}
}
\description{
Função que facilita o cálculo de milhares e milhõess
}
