% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_km.R
\name{grafico_km}
\alias{grafico_km}
\title{Grafico Kaplan-meier de modelo de sobrevivencia}
\usage{
grafico_km(m)
}
\arguments{
\item{m}{modelo}
}
\description{
funcao que cria grafico no modelo de sobrevivendia Kaplan-meier
}
