% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_km2.R
\name{grafico_km2}
\alias{grafico_km2}
\title{Gráfico Kaplan-meier de modelo de sobrevivência (2 variáveis)}
\usage{
grafico_km2(m, lab)
}
\arguments{
\item{m}{modelo}

\item{lab}{label para cores}
}
\description{
Função que cria gráfico no modelo de sobrevivência Kaplan-meier
}
